return log likelihood and average of p.
The function will has a new argument. If only.loglik=TRUE then -logL is
returned. This is what should be used for maximizing the likelihood for
a single closed population.

If only.loglik = FALSE then a list is returned: logL and the average of
the pdot values. This is used for when this closed likelihood is part of
the full likelihood for a spatial robust design.