log likelihood was producing NaN
The total hazard (h.) is producing 0 (or very near 0) values for all
traps for some of the mask locations.  p_ks is created, in part, from
the division of h. which produces NaN values.

These NaN values are then also in p.omega which is used to calculate the
likelihood.

When the sum is taken across p.omega I remove the NaN values. I'm not
sure if this is correct. I need to talk with Trent about this.